+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.0.15-alpha
+* Added timestamp for tail tacking
+
+### 0.0.14-alpha
+* Fixed issue for missing partitions in progress handler.
+
 ### 0.0.13-alpha
 * lower FSharp.Core version requirements